requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.7
Django==3.2.12
djangorestframework==3.13.1
importlib-metadata==4.8.3
kombu==4.6.11
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1
vine==1.3.0
zipp==3.6.0
